+++ actors/deposit-screener/src/screening/checks/RuleSet.ts
@@ -7,6 +7,7 @@ import {
   ApiCallReturnData,
 } from "./apiCalls";
 import IORedis from "ioredis";
+import { Logger } from "winston";
 
 export interface Rejection {
   type: "Rejection";
@@ -157,13 +158,16 @@ export class RuleSet {
   private tail: RuleLike | null = null;
   private readonly baseDelaySeconds;
   private cache: IORedis;
+  private logger: Logger;
 
   constructor(
     { baseDelaySeconds }: { baseDelaySeconds: number },
-    cache: IORedis
+    cache: IORedis,
+    logger: Logger
   ) {
     this.baseDelaySeconds = baseDelaySeconds;
     this.cache = cache;
+    this.logger = logger;
   }
 
   private _add(ruleLike: RuleLike) {
@@ -207,7 +211,11 @@ export class RuleSet {
         result,
       });
       if (result.type === "Rejection") {
-        console.log(`Screener execution for deposit:`, deposit, rulesLogList);
+        this.logger.info(
+          `Screener execution for deposit:`,
+          deposit,
+          rulesLogList
+        );
         return result;
       } else if (result.type === "Delay") {
         delaySeconds = APPLY_DELAY_OPERATION[result.operation](
@@ -217,7 +225,7 @@ export class RuleSet {
       }
       currRule = currRule.next;
     }
-    console.log(`Screener execution for deposit:`, deposit, rulesLogList);
+    this.logger.info(`Screener execution for deposit:`, deposit, rulesLogList);
     return { type: "Delay", timeSeconds: delaySeconds };
   }
 }