+++ packages/frontend-sdk/src/utils.ts
@@ -23,6 +23,15 @@ import {
 } from "./types";
 
 const ENDPOINTS = {
+  mainnet: {
+    screenerEndpoint:
+      process.env.NEXT_PUBLIC_SCREENER_URL ??
+      "https://screener.mainnet.nocturnelabs.xyz",
+    bundlerEndpoint:
+      process.env.NEXT_PUBLIC_BUNDLER_URL ??
+      "https://bundler.mainnet.nocturnelabs.xyz",
+    subgraphEndpoint: process.env.NEXT_PUBLIC_SUBGRAPH_URL,
+  },
   goerli: {
     screenerEndpoint:
       process.env.NEXT_PUBLIC_SCREENER_URL ??
@@ -30,9 +39,7 @@ const ENDPOINTS = {
     bundlerEndpoint:
       process.env.NEXT_PUBLIC_BUNDLER_URL ??
       "https://bundler.testnet.nocturnelabs.xyz",
-    subgraphEndpoint:
-      process.env.NEXT_PUBLIC_SUBGRAPH_URL ??
-      "https://api.goldsky.com/api/public/project_cldkt6zd6wci33swq4jkh6x2w/subgraphs/nocturne/0.1.25-testnet/gn",
+    subgraphEndpoint: process.env.NEXT_PUBLIC_SUBGRAPH_URL,
   },
   localhost: {
     screenerEndpoint: "http://localhost:3001",
@@ -44,7 +51,7 @@ const ENDPOINTS = {
 export function getTokenContract(
   assetType: AssetType,
   assetAddress: Address,
-  signerOrProvider: ethers.Signer | ethers.providers.Provider
+  signerOrProvider: ethers.Signer | ethers.providers.Provider,
 ): ethers.Contract {
   let abi;
   if (assetType == AssetType.ERC20) {
@@ -63,13 +70,13 @@ export function getTokenContract(
 export async function getTokenDetails(
   assetType: AssetType,
   assetAddress: Address,
-  signerOrProvider: ethers.Signer | ethers.providers.Provider
+  signerOrProvider: ethers.Signer | ethers.providers.Provider,
 ): Promise<TokenDetails> {
   console.log("getting token contract...");
   const tokenContract = getTokenContract(
     assetType,
     assetAddress,
-    signerOrProvider
+    signerOrProvider,
   );
 
   if (assetType == AssetType.ERC20) {
@@ -88,14 +95,14 @@ export async function getTokenDetails(
 
 export function formatTokenAmountUserRepr(
   balance: bigint,
-  decimals: number
+  decimals: number,
 ): number {
   return Number(balance) / Math.pow(10, decimals);
 }
 
 export function formatTokenAmountEvmRepr(
   amount: number,
-  decimals: number
+  decimals: number,
 ): bigint {
   return BigInt(amount * Math.pow(10, decimals));
 }
@@ -120,7 +127,7 @@ export interface CircuitArtifactUrls {
 }
 
 export function getCircuitArtifactUrls(
-  networkName: SupportedNetwork
+  networkName: SupportedNetwork,
 ): CircuitArtifactUrls {
   switch (networkName) {
     case "mainnet":
@@ -168,14 +175,29 @@ export function getCircuitArtifactUrls(
 }
 
 export function getNocturneSdkConfig(
-  networkName: SupportedNetwork
+  networkName: SupportedNetwork,
 ): NocturneSdkConfig {
   const config = loadNocturneConfigBuiltin(networkName);
 
   let endpoints: Endpoints;
   switch (networkName) {
+    case "mainnet":
+      if (!ENDPOINTS.mainnet.subgraphEndpoint) {
+        throw new Error(
+          `Missing subgraph endpoint for network: ${networkName}`,
+        );
+      }
+
+      endpoints = ENDPOINTS.mainnet as Endpoints;
+      break;
     case "goerli":
-      endpoints = ENDPOINTS.goerli;
+      if (!ENDPOINTS.goerli.subgraphEndpoint) {
+        throw new Error(
+          `Missing subgraph endpoint for network: ${networkName}`,
+        );
+      }
+
+      endpoints = ENDPOINTS.goerli as Endpoints;
       break;
     case "localhost":
       endpoints = ENDPOINTS.localhost;
@@ -191,7 +213,7 @@ export function getNocturneSdkConfig(
 }
 
 export function toDepositRequest(
-  displayDepositRequest: DisplayDepositRequest
+  displayDepositRequest: DisplayDepositRequest,
 ): DepositRequest {
   const asset = {
     ...displayDepositRequest.asset,
@@ -212,7 +234,7 @@ export function toDepositRequest(
 
 export function flattenDepositRequestStatus(
   subgraphStatus: OnChainDepositRequestStatus,
-  screenerStatus: ScreenerDepositRequestStatus
+  screenerStatus: ScreenerDepositRequestStatus,
 ): DepositRequestStatus {
   switch (subgraphStatus) {
     case OnChainDepositRequestStatus.Retrieved:
@@ -240,9 +262,10 @@ export function flattenDepositRequestStatus(
 }
 
 // TODO: flatten these two option types and change all tests to expect new behavior
-export function getBalanceOptsToGetNotesOpts(
-  { includeUncommitted, includePending }: GetBalanceOpts
-): GetNotesOpts {
+export function getBalanceOptsToGetNotesOpts({
+  includeUncommitted,
+  includePending,
+}: GetBalanceOpts): GetNotesOpts {
   return {
     includeUncommitted,
     ignoreOptimisticNFs: !includePending,