+++ actors/balance-monitor/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [8742f9a0]
+  - @nocturne-xyz/config@1.6.0
+
 ## 0.3.0
 
 ### Minor Changes